Move emptystavesvisibilitymodel to separate file
