+++ src/notation/view/internal/staffvisibilitypopupmodel.h
@@ -25,12 +25,8 @@
 #include <QAbstractItemModel>
 #include <QQmlParserStatus>
 #include <memory>
-#include <vector>
 
-#include "context/iglobalcontext.h"
-#include "inotation.h"
-#include "modularity/ioc.h"
-#include "types/types.h"
+#include "emptystavesvisibilitymodel.h"
 #include "view/abstractelementpopupmodel.h"
 
 namespace mu::engraving {
@@ -65,63 +61,5 @@ private:
 
     std::unique_ptr<EmptyStavesVisibilityModel> m_emptyStavesVisibilityModel = nullptr;
     size_t m_systemIndex = 0;
-};
-
-class EmptyStavesVisibilityModel : public QAbstractItemModel, public muse::Injectable
-{
-    Q_OBJECT
-
-    Q_PROPERTY(bool canResetAll READ canResetAll NOTIFY canResetAllChanged)
-
-    muse::Inject<context::IGlobalContext> globalContext { this };
-
-public:
-    explicit EmptyStavesVisibilityModel(QObject* parent = nullptr);
-    ~EmptyStavesVisibilityModel() override;
-
-    void load(INotationPtr notation, engraving::System* system);
-
-    QModelIndex index(int row, int column, const QModelIndex& parent = QModelIndex()) const override;
-    QModelIndex parent(const QModelIndex& child) const override;
-    int rowCount(const QModelIndex& parent = QModelIndex()) const override;
-    int columnCount(const QModelIndex& parent = QModelIndex()) const override;
-    QVariant data(const QModelIndex& index, int role = Qt::DisplayRole) const override;
-    bool setData(const QModelIndex& index, const QVariant& value, int role = Qt::EditRole) override;
-    Q_INVOKABLE void resetVisibility(const QModelIndex& index);
-    Q_INVOKABLE void resetAllVisibility();
-    QHash<int, QByteArray> roleNames() const override;
-
-    int startSystemIndex() const;
-
-signals:
-    void canResetAllChanged();
-
-private:
-    enum Roles {
-        Name = Qt::UserRole + 1,
-        IsVisible,
-        CanChangeVisibility,
-        CanReset
-    };
-
-    struct Item;
-    struct PartItem;
-    struct StaffItem;
-
-    void reload();
-
-    bool canResetAll() const;
-
-    void setPartVisibility(PartItem* partItem, engraving::AutoOnOff value);
-    void setStaffVisibility(StaffItem* staffItem, engraving::AutoOnOff value);
-
-    void updateData(PartItem* partItem);
-
-    int partIndex(const PartItem* partItem) const;
-
-    INotationPtr m_notation;
-    engraving::System* m_system = nullptr;
-
-    std::vector<std::unique_ptr<PartItem> > m_parts;
 };
 }
